refactor: reorganize resource ROI tests
