+++ tests/test_resource_rois.py
@@ -408,109 +408,6 @@ class TestResourceROIs(TestCase):
         self.assertEqual(width2, expected_width2, "width not limited by max_width[1]")
         self.assertEqual(left2, avail_left2)
 
-    def test_build_rois_reduce_width_when_gap_below_min(self):
-        """Icons closer than ``min_width`` should yield a shrunken ROI within the gap."""
-
-        detected = {
-            "wood_stockpile": (0, 0, 5, 5),
-            "food_stockpile": (15, 0, 5, 5),
-        }
-        regions, spans, narrow = resources.compute_resource_rois(
-            0,
-            100,
-            0,
-            10,
-            [2] * 6,
-            [2] * 6,
-            [0] * 6,
-            [999] * 6,
-            [20] * 6,
-            detected=detected,
-        )
-        roi = regions["wood_stockpile"]
-        span_left = 7  # icon_right + pad_left
-        span_right = 13  # next_icon_left - pad_right
-        self.assertEqual(roi[0], span_left)
-        self.assertEqual(roi[0] + roi[2], span_right)
-        self.assertEqual(roi[2], span_right - span_left)
-        self.assertIn("wood_stockpile", narrow)
-        self.assertEqual(narrow["wood_stockpile"], 14)
-
-    def test_ignores_non_positive_span(self):
-        """Resources with zero or negative spans should be skipped entirely."""
-
-        detected = {
-            "wood_stockpile": (0, 0, 5, 5),
-            "food_stockpile": (6, 0, 5, 5),
-        }
-        regions, spans, narrow = resources.compute_resource_rois(
-            0,
-            100,
-            0,
-            10,
-            [2] * 6,
-            [2] * 6,
-            [0] * 6,
-            [999] * 6,
-            [0] * 6,
-            detected=detected,
-        )
-        self.assertNotIn("wood_stockpile", regions)
-        self.assertNotIn("wood_stockpile", spans)
-        self.assertNotIn("wood_stockpile", narrow)
-        self.assertIn("food_stockpile", spans)
-
-    def test_per_icon_min_width(self):
-        frame = np.zeros((50, 100, 3), dtype=np.uint8)
-        panel_box = (0, 0, 100, 20)
-
-        icons = ["wood_stockpile", "food_stockpile", "gold_stockpile"]
-        positions = [0, 39, 78]
-        pad_left = 2
-        pad_right = 2
-        min_widths = [40, 10, 0, 0, 0, 0]
-        loc_iter = iter([(x, 0) for x in positions])
-
-        def fake_minmax(res):
-            xi, yi = next(loc_iter)
-            return 0.0, 0.95, (0, 0), (xi, yi)
-
-        with patch("script.resources.find_template", return_value=(panel_box, 0.9, None)), \
-            patch("script.resources.cv2.cvtColor", lambda src, code: np.zeros(src.shape[:2], dtype=np.uint8)), \
-            patch("script.resources.cv2.resize", lambda img, *a, **k: img), \
-            patch("script.resources.cv2.matchTemplate", lambda *a, **k: np.zeros((100, 200), dtype=np.float32)), \
-            patch("script.resources.cv2.minMaxLoc", side_effect=fake_minmax), \
-            patch.object(screen_utils, "_load_icon_templates", lambda: None), \
-            patch.object(screen_utils, "HUD_TEMPLATE", np.zeros((1, 1), dtype=np.uint8)), \
-            patch.dict(screen_utils.ICON_TEMPLATES, {name: np.zeros((5, 5), dtype=np.uint8) for name in icons}, clear=True), \
-            patch.dict(common.CFG["resource_panel"], {
-                "roi_padding_left": [pad_left] * 6,
-                "roi_padding_right": [pad_right] * 6,
-                "icon_trim_pct": [0] * 6,
-                "scales": [1.0],
-                "match_threshold": 0.5,
-                "max_width": 999,
-                "min_width": min_widths,
-                "top_pct": 0.0,
-                "height_pct": 1.0,
-            }), patch.dict(
-                common.CFG["profiles"]["aoe1de"]["resource_panel"],
-                {"icon_trim_pct": [0] * 6},
-            ):
-                resources.cache._NARROW_ROIS.clear()
-                resources.cache._NARROW_ROI_DEFICITS.clear()
-                resources.RESOURCE_CACHE.resource_failure_counts.clear()
-                resources.RESOURCE_CACHE.last_resource_values.clear()
-                resources.RESOURCE_CACHE.last_resource_ts.clear()
-                resources.locate_resource_panel(frame)
-                result = resources.cache._NARROW_ROIS.copy()
-                resources.cache._NARROW_ROIS.clear()
-                resources.cache._NARROW_ROI_DEFICITS.clear()
-                resources.RESOURCE_CACHE.resource_failure_counts.clear()
-                resources.RESOURCE_CACHE.last_resource_values.clear()
-                resources.RESOURCE_CACHE.last_resource_ts.clear()
-
-        self.assertEqual(result, {"wood_stockpile"})
 
     def test_cache_cleared_on_region_change(self):
         resources.RESOURCE_CACHE.last_icon_bounds.clear()
